requirements.txt
BeautifulSoup4
requests
lxml
pprint
selenium
mysql-connector-python
python-slugify